Major refactor - reduce pages
