+++ templates/play/album-choice.pt
