--- templates/play/play-album.pt
@@ -1,14 +1,0 @@
-<div metal:use-macro="load: ../shared/_layout.pt">
-    <div metal:fill-slot="content" tal:omit-tag="True">
-
-    <div style="height 100%; width 100%;">
-        <div class="container-fluid bg-dark pt-3 d-flex align-items-center flex-column text-white min-vh-75">
-            <h2>Play an Album</h2>
-            <p>
-                Here is your random album.
-            </p>
-
-        </div>
-    </div>
-</div>
-</div>
