+++ viewmodels/play/now_playing_viewmodel.py
@@ -6,7 +6,7 @@ from services.play_service import RandomRecordService
 from viewmodels.shared.viewmodel import ViewModelBase
 
 
-class PlayAlbumViewModel(ViewModelBase):
+class NowPlayingViewModel(ViewModelBase):
     def __init__(self, request: Request):
         super().__init__(request)
 
@@ -16,8 +16,8 @@ class PlayAlbumViewModel(ViewModelBase):
         self.artist_url: str = RandomRecordService.get_album_data()
         self.release_image_uri: str = RandomRecordService.get_album_data()
         self.release_title: str = RandomRecordService.get_album_data()
-        self.release_data: str = RandomRecordService.get_album_data()
-        self.release_data: int = RandomRecordService.get_album_data()
+        self.release_date: str = RandomRecordService.get_album_data()
+        self.release_date: int = RandomRecordService.get_album_data()
         self.genres: str = RandomRecordService.get_album_data()
         self.discogs_main_id: int = RandomRecordService.get_album_data()
         self.discogs_main_url: str = RandomRecordService.get_album_data()
